# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Les Casaniers API",
        "description": "Documentation de l'API pour le projet Les Casaniers",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur Principal"
        }
    ],
    "paths": {
        "/admin/register": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Creer un compte administrateur",
                "description": "Permet de creer un nouveau compte administrateur.",
                "operationId": "4c9ff7f9fab7f89d744513cfe5480a0f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prenom",
                                    "nom",
                                    "email",
                                    "mot_de_passe"
                                ],
                                "properties": {
                                    "prenom": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@lescasaniers.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "mot_de_passe": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Secret123!"
                                    },
                                    "mot_de_passe_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Secret123!"
                                    },
                                    "poste": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "support",
                                            "logistique"
                                        ],
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte cree avec succes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte administrateur cree avec succes"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur serveur"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Connexion administrateur",
                "description": "Permet a un administrateur de se connecter via Sanctum stateful.",
                "operationId": "ca85837b6fc19d94a39a445d949d4ec9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "mot_de_passe"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@lescasaniers.com"
                                    },
                                    "mot_de_passe": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Secret123!"
                                    },
                                    "remember": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion reussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion reussie"
                                        },
                                        "data": {
                                            "properties": {
                                                "admin": {
                                                    "type": "object"
                                                },
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "1|Abc..."
                                                },
                                                "refresh_token": {
                                                    "type": "string",
                                                    "example": "2|Xyz..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 1800
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/admin/logout": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Deconnexion administrateur",
                "description": "Deconnecte l'administrateur et revoque le token.",
                "operationId": "ca3cbc352eab76d0459b48f3778c6b28",
                "responses": {
                    "200": {
                        "description": "Deconnexion reussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Deconnexion reussie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/refresh-token": {
            "post": {
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Renouveler l'access token via le refresh token",
                "operationId": "33a0fb71e1dfe60c51a79595edf92994",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refresh_token"
                                ],
                                "properties": {
                                    "refresh_token": {
                                        "type": "string",
                                        "example": "2|Xyz..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token renouvelé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "3|NewAccess..."
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "example": 1800
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refresh token invalide ou expiré"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}